+++ desktop-app/message.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"github.com/asticode/go-astilectron"
+	"github.com/asticode/go-astilectron-bootstrap"
+
+)
+
+func handleMessages(_ *astilectron.Window, m bootstrap.MessageIn) (payload interface{}, err error) {
+	return
+}