+++ src/app.ts
@@ -1,13 +1,14 @@
-import { Map } from './models/map.js'
-import { View } from './views/view.js'
-import { Dispatcher } from './dispatcher.js'
-import { AppEvent, MouseMode, Values } from './enums/enums.js'
-import { Editor } from './editor.js'
-import { Tabs } from './controls/controls.js'
-import { BlockPopup, ConnectorPopup, NotePopup, RoomPopup } from './popups/popups.js'
-import { BlockPanel, ConnectorPanel, RenderPanel, MapPanel, MenuPanel, NotePanel, RoomPanel, ToolPanel } from './panels/panels.js'
-import { MapJSON } from './io/mapJSON.js'
-import { Selection } from './selection.js'
+import { Map } from './models/map.js';
+import { View } from './views/view.js';
+import { Dispatcher } from './dispatcher.js';
+import { AppEvent, MouseMode, Values } from './enums/enums.js';
+import { Editor } from './editor.js';
+import { Tabs } from './controls/controls.js';
+import { BlockPopup, ConnectorPopup, NotePopup, RoomPopup } from './popups/popups.js';
+import { BlockPanel, ConnectorPanel, RenderPanel, MapPanel, MenuPanel, NotePanel, RoomPanel, ToolPanel } from './panels/panels.js';
+import { MapJSON } from './io/mapJSON.js';
+import { Selection } from './selection.js';
+import { Desktop } from './desktop.js';
 
 export class App {
   // - App holds the current map.
@@ -32,6 +33,10 @@ export class App {
     App.createPanels();
 
     Tabs.initialize();
+
+    document.addEventListener('astilectron-ready', function () {
+      Desktop.initialize();
+    })
   }
 
   static createPanels() {