+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to "Yggdrasil" will be documented in this file.
 
+## [0.1.0] - 2019-08-09
+
+#### Added
+* Initial desktop client.
+  * Electron ([Go-Astilectron](https://github.com/asticode/go-astilectron)) based.
+  * Windows, Mac OS and GNU/Linux support.
+  * Quickly choose from a set of predefined sample maps.
+
+  ![](changes/0.1.0/maps.png)
+
 ## [0.0.2] - 2019-08-08
 
 #### Improvement